gvta init started

+++ cpu/PTA.h
@@ -71,11 +71,11 @@ class PTA : public AA<T,Z>{
 		}
 
 		void init();
-		void findTopKscalar(uint64_t k,uint8_t qq, T *weights, uint8_t *attr);
-		void findTopKsimd(uint64_t k,uint8_t qq, T *weights, uint8_t *attr);
-		void findTopKthreads(uint64_t k,uint8_t qq, T *weights, uint8_t *attr);
-		void findTopKthreads2(uint64_t k,uint8_t qq, T *weights, uint8_t *attr);
-		void findTopKsimdMQ(uint64_t k,uint8_t qq, T *weights, uint8_t *attr, uint32_t tid);
+		void findTopKscalar(uint64_t k,uint8_t qq, T *weights, uint32_t *attr);
+		void findTopKsimd(uint64_t k,uint8_t qq, T *weights, uint32_t *attr);
+		void findTopKthreads(uint64_t k,uint8_t qq, T *weights, uint32_t *attr);
+		void findTopKthreads2(uint64_t k,uint8_t qq, T *weights, uint32_t *attr);
+		void findTopKsimdMQ(uint64_t k,uint8_t qq, T *weights, uint32_t *attr, uint32_t tid);
 
 	private:
 		pta_partition<T,Z> parts[PPARTITIONS];
@@ -285,7 +285,7 @@ void PTA<T,Z>::init(){
 }
 
 template<class T, class Z>
-void PTA<T,Z>::findTopKscalar(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
+void PTA<T,Z>::findTopKscalar(uint64_t k, uint8_t qq, T *weights, uint32_t *attr){
 	std::cout << this->algo << " find top-" << k << " scalar (" << (int)qq << "D) ...";
 	if(STATS_EFF) this->tuple_count = 0;
 	if(STATS_EFF) this->pop_count=0;
@@ -362,7 +362,7 @@ void PTA<T,Z>::findTopKscalar(uint64_t k, uint8_t qq, T *weights, uint8_t *attr)
 }
 
 template<class T, class Z>
-void PTA<T,Z>::findTopKsimd(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
+void PTA<T,Z>::findTopKsimd(uint64_t k, uint8_t qq, T *weights, uint32_t *attr){
 	std::cout << this->algo << " find top-" << k << " simd (" << (int)qq << "D) ...";
 	if(STATS_EFF) this->tuple_count = 0;
 	if(STATS_EFF) this->pop_count=0;
@@ -429,7 +429,7 @@ void PTA<T,Z>::findTopKsimd(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
 }
 
 template<class T, class Z>
-void PTA<T,Z>::findTopKsimdMQ(uint64_t k, uint8_t qq, T *weights, uint8_t *attr, uint32_t tid){
+void PTA<T,Z>::findTopKsimdMQ(uint64_t k, uint8_t qq, T *weights, uint32_t *attr, uint32_t tid){
 	Time<msecs> t;
 	if(STATS_EFF) this->tuple_count = 0;
 	if(STATS_EFF) this->pop_count=0;
@@ -484,7 +484,7 @@ void PTA<T,Z>::findTopKsimdMQ(uint64_t k, uint8_t qq, T *weights, uint8_t *attr,
 }
 
 template<class T, class Z>
-void PTA<T,Z>::findTopKthreads(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
+void PTA<T,Z>::findTopKthreads(uint64_t k, uint8_t qq, T *weights, uint32_t *attr){
 	uint32_t threads = THREADS < PPARTITIONS ? THREADS : PPARTITIONS;
 	Z tt_count[threads];
 	std::priority_queue<T, std::vector<tuple_<T,Z>>, MaxCMP<T,Z>> _q[threads];
@@ -573,7 +573,7 @@ void PTA<T,Z>::findTopKthreads(uint64_t k, uint8_t qq, T *weights, uint8_t *attr
 }
 
 template<class T, class Z>
-void PTA<T,Z>::findTopKthreads2(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
+void PTA<T,Z>::findTopKthreads2(uint64_t k, uint8_t qq, T *weights, uint32_t *attr){
 	uint32_t threads = THREADS < PPARTITIONS ? THREADS : PPARTITIONS;
 	Z tt_count[threads];
 	std::priority_queue<T, std::vector<tuple_<T,Z>>, MaxCMP<T,Z>> _q[threads];
