+++ cpu/VTA.h
@@ -60,11 +60,11 @@ class VTA : public AA<T,Z>{
 
 		}
 		void init();
-		void findTopKscalar(uint64_t k,uint8_t qq, T *weights, uint8_t *attr);
-		void findTopKsimd(uint64_t k,uint8_t qq, T *weights, uint8_t *attr);
-		void findTopKthreads(uint64_t k,uint8_t qq, T *weights, uint8_t *attr);
-		void findTopKthreads2(uint64_t k,uint8_t qq, T *weights, uint8_t *attr);
-		void findTopKsimdMQ(uint64_t k,uint8_t qq, T *weights, uint8_t *attr, uint32_t tid);
+		void findTopKscalar(uint64_t k,uint8_t qq, T *weights, uint32_t *attr);
+		void findTopKsimd(uint64_t k,uint8_t qq, T *weights, uint32_t *attr);
+		void findTopKthreads(uint64_t k,uint8_t qq, T *weights, uint32_t *attr);
+		void findTopKthreads2(uint64_t k,uint8_t qq, T *weights, uint32_t *attr);
+		void findTopKsimdMQ(uint64_t k,uint8_t qq, T *weights, uint32_t *attr, uint32_t tid);
 
 	private:
 		vta_partition<T,Z> parts[VPARTITIONS];
@@ -158,7 +158,7 @@ void VTA<T,Z>::init(){
 }
 
 template<class T, class Z>
-void VTA<T,Z>::findTopKscalar(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
+void VTA<T,Z>::findTopKscalar(uint64_t k, uint8_t qq, T *weights, uint32_t *attr){
 	std::cout << this->algo << " find top-" << k << " scalar (" << (int)qq << "D) ...";
 	if(STATS_EFF) this->tuple_count = 0;
 	if(STATS_EFF) this->pop_count=0;
@@ -231,7 +231,7 @@ void VTA<T,Z>::findTopKscalar(uint64_t k, uint8_t qq, T *weights, uint8_t *attr)
 }
 
 template<class T, class Z>
-void VTA<T,Z>::findTopKsimd(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
+void VTA<T,Z>::findTopKsimd(uint64_t k, uint8_t qq, T *weights, uint32_t *attr){
 	std::cout << this->algo << " find top-" << k << " simd (" << (int)qq << "D) ...";
 	if(STATS_EFF) this->tuple_count = 0;
 	if(STATS_EFF) this->pop_count=0;
@@ -339,7 +339,7 @@ void VTA<T,Z>::findTopKsimd(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
 }
 
 template<class T, class Z>
-void VTA<T,Z>::findTopKsimdMQ(uint64_t k, uint8_t qq, T *weights, uint8_t *attr, uint32_t tid){
+void VTA<T,Z>::findTopKsimdMQ(uint64_t k, uint8_t qq, T *weights, uint32_t *attr, uint32_t tid){
 	Time<msecs> t;
 	if(STATS_EFF) this->tuple_count = 0;
 	if(STATS_EFF) this->pop_count=0;
@@ -420,7 +420,7 @@ void VTA<T,Z>::findTopKsimdMQ(uint64_t k, uint8_t qq, T *weights, uint8_t *attr,
 }
 
 template<class T, class Z>
-void VTA<T,Z>::findTopKthreads(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
+void VTA<T,Z>::findTopKthreads(uint64_t k, uint8_t qq, T *weights, uint32_t *attr){
 	uint32_t threads = THREADS < VPARTITIONS ? THREADS : VPARTITIONS;
 	Z tt_count[threads];
 	std::priority_queue<T, std::vector<tuple_<T,Z>>, PQComparison<T,Z>> q[threads];
@@ -508,7 +508,7 @@ void VTA<T,Z>::findTopKthreads(uint64_t k, uint8_t qq, T *weights, uint8_t *attr
 }
 
 template<class T, class Z>
-void VTA<T,Z>::findTopKthreads2(uint64_t k, uint8_t qq, T *weights, uint8_t *attr){
+void VTA<T,Z>::findTopKthreads2(uint64_t k, uint8_t qq, T *weights, uint32_t *attr){
 	uint32_t threads = THREADS;
 	Z tt_count[threads];
 	std::priority_queue<T, std::vector<tuple_<T,Z>>, PQComparison<T,Z>> q[threads];
